Update LICENSE
Feed the Charity to Feed the Homeless
Feed the Charity to Feed the Homeless is a purpose-driven NFT collection aimed at addressing homelessness by leveraging blockchain technology to fund charitable initiatives. This project ties Ricky Amaya’s creative expertise and philanthropic vision with the power of NFTs to create a meaningful and impactful movement.